+++ packages/scorers/src/migrations/index.ts
@@ -8,11 +8,13 @@ import * as Migrator from "effect/unstable/sql/Migrator"
 import scores from "./0001_scores.ts"
 import jobs from "./0002_score_jobs.ts"
 import failureCodes from "./0003_score_failure_codes.ts"
+import requiredFailureCodes from "./0004_require_failure_codes.ts"
 
 const migrations = {
   "0001_scores": scores,
   "0002_score_jobs": jobs,
-  "0003_score_failure_codes": failureCodes
+  "0003_score_failure_codes": failureCodes,
+  "0004_require_failure_codes": requiredFailureCodes
 }
 
 /**